backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
sqlalchemy>=1.4
pydantic>=1.10,<2.0
python-jose>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6
python-dotenv>=1.0
email-validator>=2.0
bcrypt<4
